Starry Sky

+++ include/block.h
@@ -97,7 +97,7 @@ public:
     Block(float X = 0, float Y = 0, uint8_t t = 0, uint8_t l = 0, const char *P = ""); 
 
     // Init grid and other stuff for normal block
-    void init(uint16_t sf);
+    void init(float sf);
 
     // Setters
     void setOnScreen(bool os);
